track labels through product, part 1
